+++ dll/dll/steam_app_ids.h
@@ -1,0 +1,8 @@
+#ifndef _STEAM_APP_IDS_H_
+#define _STEAM_APP_IDS_H_
+
+#include "common_includes.h"
+
+extern const std::map<uint32, std::string> steam_preowned_app_ids;
+
+#endif // _STEAM_APP_IDS_H_